tools/requirements.txt
openai>=1.0.0
edge-tts>=6.1.0
pydub>=0.25.1
requests>=2.31.0
Pillow>=10.0.0
pyautogui>=0.9.54
pyperclip>=1.8.2

